requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-modeltranslation==0.18.4
django-timezone-field==5.0
executing==0.10.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.13.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
